push before writing code for loading the ball speed

+++ Assets/Activity 1 - Ball and Balloons/Scripts/ActivityOneSave.cs
@@ -9,7 +9,7 @@ public class ActivityOneSave
 
     public float BallSpeed;
 
-    public float BallSize;
+    public int BallSize;
 
     public int[] Grid = new int[2];
